Add initial training log for PPO with scene pool and environment details
- Created a new log file `temp-reset-tensor-smoke.log` to capture the training process.
- Logged initialization of unified logging for the `navi_actor_train`.
- Included details about the scene pool, loaded assets, and selected spawn candidates.
- Documented the training process, including CUDA preflight checks, memory usage, and training steps.
- Captured warnings related to soft stall monitoring and performance metrics during training.
- Logged completion of inline PPO optimization and saving of training checkpoints.

+++ projects/auditor/src/navi_auditor/dashboard/renderers.py
@@ -326,7 +326,7 @@ def render_first_person(
         return img, VIEW_RANGE_M
 
     # Pitch quantisation for remap cache (centidegrees)
-    pitch_q = int(round(float(np.rad2deg(pitch)) * 100.0))
+    pitch_q = round(float(np.rad2deg(pitch)) * 100.0)
     flen = view_w / (2.0 * np.tan(np.deg2rad(MAIN_FOV_DEG) / 2.0))
     horizon_y = int(np.clip(height * 0.54 + pitch * flen, 0, height - 1))
